fix(state): eliminate the salience/outcome repository import cycle (#672)
bun run test:unit caught a real regression the first refactor commit
introduced: tests/architecture/import-cycle-ratchet.test.ts enforces a
zero-tolerance, shrink-only baseline (CYCLE_PARTICIPANT_BASELINE is empty)
over the whole src/ static import graph, and it counts type-only imports
as graph edges same as value imports (dependency direction is treated as
an architecture property, not a runtime one). The initial repository
files each imported a type back from the domain module they re-export
from -- salience-repository.ts imported EncodingSource/SalienceVector
from commands/improve/salience.ts, and outcome-repository.ts imported
AssetOutcomeRow from commands/improve/outcome-loop.ts -- while the
domain modules import the repositories' functions for re-export. Two
2-node cycles. The baseline cannot be widened to admit them (the ratchet
is deliberately absolute; adding entries is itself a flagged violation of
the "never grows" test).

Fix: make both repository files import NOTHING from the domain modules
they serve, so the mandatory domain -> repository edge (for re-export)
stays one-directional:
- salience-repository.ts: AssetSalienceRow.encoding_source now uses the
  inline literal union "content" | "type-stub" | null (structurally
  identical to EncodingSource) instead of importing the type; a local
  SalienceVectorLike interface (fields upsertAssetSalience actually
  reads) replaces the SalienceVector import -- any real SalienceVector
  value is still assignable, so callers are unaffected.
- outcome-repository.ts: AssetOutcomeRow (plain string/number fields,
  no cross-type dependency) now lives here instead of in
  outcome-loop.ts, which imports it back and re-exports it -- this mirrors
  the AssetSalienceRow precedent (already living in its repository)
  exactly, and is a more consistent design than the original split.

Behavior is unchanged: both are pure type-shape adjustments (structural
mirrors / relocations), not logic changes. The CASE guards, SQL text, and
bind order from the prior commit are untouched.

Verified: bun scripts/lint-import-cycles.ts reports "0 cycle
participant(s)"; bunx tsc --noEmit clean; the same 168 pure-refactor +
lint-meta tests from the prior two commits still pass unchanged; full
bun run test:unit (2986 pass / 0 fail) and bun run test:integration
(5095 pass / 0 fail) both green.
